Add unit tests for duplicating a shared note and fix some related edge cases.
I can't tell if ShortcutUtils.anonymizeStyleShortcuts() is efficient or really slow. I think it's generally about O(n^2), but in one edge case, it's O(n^3)... Regardless, the goal of that is to make it more efficient for the frontend to do its thing, as it'll be doing it more often.